Open Source would assume that it is a full open source project
We want to word things, the project is open but to contribution. People taking this product and using it for their own projects would be a bit of a hurtful thing since the code has been initiated here. So projects will be open source. I think we should word carefully on the side of caution in attempt to inform contributors.